LayoutTests/imported/w3c/web-platform-tests/domxpath/fn-substring-after-expected.txt

^^^bar$$$bar^bar

FAIL substring-after() arguments depending on the context node assert_equals: expected "$$$" but got "^^bar$$$"
